spellsource-art-generation/spellsource_art_generation/requirements.txt
comfyui
diffusers>=0.16.1
bitsandbytes>=0.40.1
rembg
optimum
importlib_metadata
huggingface_hub
scipy
opencv-python>=4.8.0
filelock
numpy
Pillow
einops
timm==0.6.13
scikit-image
matplotlib
python-dateutil
mediapipe
svglib
fvcore
yapf
omegaconf
ftfy
addict